% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_TGV.R
\name{calc_TGV}
\alias{calc_TGV}
\title{Cacluate Total Genetic Value}
\usage{
calc_TGV(
  geno.info,
  map.info,
  cross.design = NULL,
  A,
  a,
  dom.coeff,
  founder = F,
  rqtl.dom = NULL,
  par.markers.unique = F
)
}
\arguments{
\item{geno.info}{The object that is returned from create parents or make_crosses}

\item{map.info}{The object returned from create_map()}

\item{cross.design}{Object returned from create_cross_design()}

\item{A}{Value assigned to the Major SNPQTL allele}

\item{a}{Value assigned to the Minor SNPQTL allele}

\item{dom.coeff}{The dominance coeffcient used for SNPQTLs}

\item{founder}{logical. Only should be set to true for founder population.}

\item{rqtl.dom}{The dominance coeffcients used for random qtl}

\item{par.markers.unqiue}{logical. Set to TRUE if create_parents was used to have unique markers}
}
\description{
This function esimates the total genetic value of parents produced from the create_progeny function
}
\examples{
parents.TGV <- calc_TGV(geno.info = parents,map.info = genetic.map,A = 1,a = -100,dom.coeff = 0,founder = T)
}
\keyword{genetic}
\keyword{progeny}
\keyword{value}
