% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_map.R
\name{create_genetic_map}
\alias{create_genetic_map}
\title{Simulate Genetic Map}
\usage{
create_genetic_map(
  num.chromos,
  map.length,
  num.markers,
  total.qtl,
  num.snpqtl,
  map.dist = "haldane",
  distribute.loci = NULL,
  marker.distribution = NULL,
  loci.per.chrom = NULL,
  marker.maf = NULL,
  snp.qtl.maf = NULL,
  chromosome.size.range = 0.2,
  save = F
)
}
\arguments{
\item{num.chromos}{The number of chromosomes to simulate}

\item{map.length}{The genetic map length in centi-morgans}

\item{num.markers}{Number of markers that can be used for genomic selection or estimated relatedness}

\item{total.qtl}{The total number of QTL to place on the genetic map}

\item{num.snpqtl}{The number of total.qtl which will be SNPs}

\item{map.dist}{Type of distance function to use in calculating recombination frequences.  Options are: "haldane" & "kosambi".  (Default = "haldane)}

\item{distribute.loci}{Default is NULL so that loci are randomly distributed to linkage groups. Options: "even" or "list"}

\item{marker.distribution}{Changes the assignment of markers to linkage groups from random to equally spaced ("equally-spaced")}

\item{marker.maf}{Default is Null. Unless specified markers MAFs will be drawn from beta distribution}

\item{snp.qtl.maf}{A vector of 2 numbers indiciating the min and max range of SNP QTL MAFs}

\item{chromosome.size.range}{The chromosome size range from the mean length of chromosomes (Default: 0.2)}

\item{save}{logical. Saves the output of genetic map (Default: FALSE)}
}
\description{
This function simulates the creation of a genetic map which mimics the genetic architecture of user defined inputs
}
\examples{
 create_genetic_map(save=T,num.chromos = 12,map.length = 1800,num.markers = 120,total.qtl = 120, num.snpqtl = 120,distribute.loci = "even",marker.distribution = "equally-spaced",snp.qtl.maf = c(0.01,0.02))
}
\keyword{cat}
