% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_selections.R
\name{extract_selections}
\alias{extract_selections}
\title{Extract Selections}
\usage{
extract_selections(
  map.info,
  cross.design,
  past.tgv,
  past.phenos,
  parent.info,
  progeny.info,
  progeny.TGV,
  progeny.phenos,
  selection.strategy,
  among.family.selection,
  within.family.selection = NULL,
  num.selections.within.family = 1,
  num.selections.among.family = NULL,
  relationship.matrix.type = rel.mat.cross,
  prefix = NULL,
  rep.num = NULL,
  reduced = F,
  weighted = F,
  num.cores = NULL
)
}
\arguments{
\item{map.info}{The number of chromosomes to simulate}

\item{cross.design}{The genetic map length in centi-morgans}

\item{past.tgv}{Number of markers that can be used for genomic selection or estimated relatedness}

\item{past.phenos}{The total number of QTL to place on the genetic map}

\item{parent.info}{The number of total.qtl which will be SNPs}

\item{progeny.info}{Type of distance function to use in calculating recombination frequences.  Options are: "haldane" & "kosambi".  (Default = "haldane)}

\item{progeny.phenos}{Default is NULL so that loci are randomly distributed to linkage groups. Options: "even" or "list"}

\item{selection.strategy}{Changes the assignment of markers to linkage groups from random to equally spaced ("equally-spaced")}

\item{among.family.selection}{A vector defining number of loci on each linkage group. Length of vector should be equal to number of linkage groups. Default is 'NULL'.}

\item{within.family.selection}{Default is Null. Unless specified markers MAFs will be drawn from beta distribution}

\item{num.selections.within.family}{The chromosome size range from the mean length of chromosomes (Default: 0.2)}

\item{num.selections.among.family}{The chromosome size range from the mean length of chromosomes (Default: 0.2)}

\item{relationship.matrix.type}{The chromosome size range from the mean length of chromosomes (Default: 0.2)}

\item{prefix}{The chromosome size range from the mean length of chromosomes (Default: 0.2)}

\item{rep.num}{The chromosome size range from the mean length of chromosomes (Default: 0.2)}

\item{reduced}{logical. Saves the output of genetic map (Default: FALSE)}

\item{weighted}{logical. Saves the output of genetic map (Default: FALSE)}

\item{num.cores}{logical. Saves the output of genetic map (Default: FALSE)}

\item{snp.qtl.maf}{A vector of 2 numbers indiciating the min and max range of SNP QTL MAFs}
}
\description{
This function extracts selections from progeny generated utilizing specified strategy.
}
\examples{
 extract_selections(save=T,num.chromos = 12,map.length = 1800,num.markers = 120,total.qtl = 120, num.snpqtl = 120,distribute.loci = "even",marker.distribution = "equally-spaced",snp.qtl.maf = c(0.01,0.02))
}
\keyword{cat}
